apps/ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "PydanticAI agent service for agentic AI spike"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai-slim[openai,anthropic]>=1.0.0",
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.7.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_service"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
